change package name, repo to main
