+++ Dockerfile
@@ -5,4 +5,4 @@ RUN go get github.com/russross/blackfriday github.com/gorilla/mux
 
 WORKDIR /srv
 
-CMD ["go", "run", "main.go"]
+CMD ["go", "run", "repo.go"]
